fix: исправлена ошибка при формировании исключения

+++ src/main/java/medvedev/com/exception/NotEnoughFundsBalanceException.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public class NotEnoughFundsBalanceException extends RuntimeException {
 
     public NotEnoughFundsBalanceException(BigDecimal balance) {
-        super(String.format("Not enough funds on the balance. Balance = %d", balance.doubleValue()));
+        super(String.format("Not enough funds on the balance. Balance = %d", balance.intValue()));
     }
 }
